add cloud storage support for aws gcp and azure

+++ BackAnd.Web.Api/Controllers/apiController.cs
@@ -1455,7 +1455,8 @@ namespace BackAnd.Web.Api.Controllers
             if (e is DeleteEventArgs)
             {
                 Durados.Cloud deleteCloud = Map.Database.Clouds.Values.Where(c => c.Id.ToString() == e.PrimaryKey).FirstOrDefault();
-                DeleteCloudFunctions(deleteCloud);
+                if(deleteCloud != null)                    
+                    DeleteCloudFunctions(deleteCloud);
                 return;
             }
